create interfaces folder to encapsulate common interfaces
