+++ src/app/_interfaces/ContentTypes.ts
@@ -1,0 +1,8 @@
+import type { Blogpost, CaseStudy, PodcastEpisode, TalksAndRoundtable } from "@/payload/payload-types";
+
+export interface ContentTypes {
+  Blogposts: Blogpost[]
+  PodcastEpisodes: PodcastEpisode[]
+  CaseStudies: CaseStudy[]
+  TalksAndRoundtables: TalksAndRoundtable[]
+}
